+++ src/fsort_template.h
@@ -14,7 +14,7 @@
 
 #define INC(X) X=STEP(X,1)
 #define DEC(X) X=STEP(X,-1)
-#define SIZE ((long)(char *)STEP((TYPE *)0,1))
+#define SIZE ((ptrdiff_t)(char *)STEP((TYPE *)0,1))
 
 #define PARENT(X) (((X)-1)>>1)
 #define CHILD1(X) (((X)<<1)+1)
@@ -45,7 +45,7 @@ static void MKNAME(_do_sort)(register TYPE *bas,
     }else{
       if(--max_recursion <= 0)
       {
-	long howmany,x,y,z;
+	ptrdiff_t howmany,x,y,z;
 	howmany=((((char *)last)-((char *)bas))/SIZE)+1;
 	if(howmany<2) return;
